third_party/WebKit/LayoutTests/gamepad/gamepad-api-expected.txt

Validates the exposed Gamepad API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.getGamepads is defined.
PASS GamepadEvent is defined.
PASS gamepads.length is 4
PASS gamepads[0] is null
PASS gamepads.item.__proto__ is Function.prototype
PASS gamepads.item(0) is null
PASS gamepad is non-null.
PASS gamepad.connected is true
PASS gamepad.__proto__ is Gamepad.prototype
PASS gamepad.id.__proto__ is String.prototype
PASS gamepad.connected.__proto__ is Boolean.prototype
PASS gamepad.index.__proto__ is Number.prototype
PASS gamepad.timestamp.__proto__ is Number.prototype
PASS gamepad.axes.__proto__ is Array.prototype
PASS gamepad.axes[0].__proto__ is Number.prototype
PASS gamepad.buttons.__proto__ is Array.prototype
PASS gamepad.buttons[0].pressed.__proto__ is Boolean.prototype
PASS gamepad.buttons[0].value.__proto__ is Number.prototype
PASS gamepad.mapping.__proto__ is String.prototype
PASS Object.isFrozen(gamepad.axes) is true
PASS Object.isFrozen(gamepad.buttons) is true
PASS successfullyParsed is true

TEST COMPLETE
